cleanup and reuse layout
we have 2 types of documents: canvas and designer

canvas is used only for the /canvas and the /preview routes, the rest is designer until we have some special case where we need a different document for a page
