+++ app/routes/dashboard.tsx
@@ -1,1 +1,8 @@
-export { meta, default } from "./designer";
+import type { MetaFunction } from "@remix-run/node";
+import { Designer } from "~/shared/documents/designer";
+
+export const meta: MetaFunction = () => {
+  return { title: "Webstudio Dashboard" };
+};
+
+export default Designer;
