+++ app/routes/login.tsx
@@ -1,8 +1,8 @@
-import type { MetaFunction } from "@remix-run/node";
+import { MetaFunction } from "@remix-run/node";
 import { Designer } from "~/shared/documents/designer";
 
 export const meta: MetaFunction = () => {
-  return { title: "Webstudio" };
+  return { title: "Webstudio Login" };
 };
 
 export default Designer;
